defeat memory renaming
Memory renaming on Ice Lake messes up some of the
pointer chasing tests: we write a value to the
stack that points to itself, and then do a pointer
chase against it to measure the load latency in
various scenarios.

With memory renaming, this load can be satisfied
with "zero" latency from the register file (taking
its value from the store) producing unpredictable
results (since the optimization doesn't always
take effect, e.g. a context switch can defeat it).

Get sane results with an mfence before the measurement
loop, which (empirically!) defeats memory renaming,
probably because it drains the store buffer so
the store never gets a chance to forward.